+++ src/components/kytos/switch/Interface.vue
@@ -78,7 +78,7 @@ export default {
     metadata: {
       type: Object,
       required: false,
-    }
+    },
   },
   data () {
     return {